Add cookbook to be chef 13 compatible
STATE:
Chef version 13.0 does not allow the use of LWRP.

FIX:
Merge the old LWRP into the custom_resource "kernel_module"
